api/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
PyMySQL>=1.1
python-dotenv>=1.0